cleanup(generator): separate golden from hand-crafted files
Move the hand-crafted files out of the
`generator/integration_tests/golden/` subdirectory, including CMake and
Bazel support files.  I think this will make it easier to work in these
files, as they will format and work as regular files.
