+++ generator/integration_tests/README.md
@@ -17,3 +17,10 @@ bazel run \
 
 The `generate-libraries` build also runs this command and highlights any changes
 using `git diff`.
+
+## Keep hand-crafted tests separate from golden files
+
+We write unit tests to verify the generated files work as expected. These tests
+are in the `tests/` subdirectory.  Do **not** insert hand-crafted code into
+the `golden/` subdirectory. That directory does not get the usual creature
+comforts, such as automatic code formatting.
