--- generator/integration_tests/golden/google_cloud_cpp_generator_golden_files.bzl
@@ -1,59 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
-
-"""Automatically generated unit tests list - DO NOT EDIT."""
-
-google_cloud_cpp_generator_golden_files = [
-    "golden_kitchen_sink_client.cc",
-    "golden_kitchen_sink_client.h",
-    "golden_kitchen_sink_connection.cc",
-    "golden_kitchen_sink_connection.h",
-    "golden_kitchen_sink_connection_idempotency_policy.cc",
-    "golden_kitchen_sink_connection_idempotency_policy.h",
-    "golden_kitchen_sink_options.h",
-    "golden_thing_admin_client.cc",
-    "golden_thing_admin_client.h",
-    "golden_thing_admin_connection.cc",
-    "golden_thing_admin_connection.h",
-    "golden_thing_admin_connection_idempotency_policy.cc",
-    "golden_thing_admin_connection_idempotency_policy.h",
-    "golden_thing_admin_options.h",
-    "internal/golden_kitchen_sink_logging_decorator.cc",
-    "internal/golden_kitchen_sink_logging_decorator.h",
-    "internal/golden_kitchen_sink_metadata_decorator.cc",
-    "internal/golden_kitchen_sink_metadata_decorator.h",
-    "internal/golden_kitchen_sink_option_defaults.cc",
-    "internal/golden_kitchen_sink_option_defaults.h",
-    "internal/golden_kitchen_sink_retry_traits.h",
-    "internal/golden_kitchen_sink_stub.cc",
-    "internal/golden_kitchen_sink_stub.h",
-    "internal/golden_kitchen_sink_stub_factory.cc",
-    "internal/golden_kitchen_sink_stub_factory.h",
-    "internal/golden_thing_admin_logging_decorator.cc",
-    "internal/golden_thing_admin_logging_decorator.h",
-    "internal/golden_thing_admin_metadata_decorator.cc",
-    "internal/golden_thing_admin_metadata_decorator.h",
-    "internal/golden_thing_admin_option_defaults.cc",
-    "internal/golden_thing_admin_option_defaults.h",
-    "internal/golden_thing_admin_retry_traits.h",
-    "internal/golden_thing_admin_stub.cc",
-    "internal/golden_thing_admin_stub.h",
-    "internal/golden_thing_admin_stub_factory.cc",
-    "internal/golden_thing_admin_stub_factory.h",
-    "internal/streaming.cc",
-    "mocks/mock_golden_kitchen_sink_connection.h",
-    "mocks/mock_golden_thing_admin_connection.h",
-]
