--- generator/integration_tests/golden/tests/.clang-format
@@ -1,2 +1,0 @@
-BasedOnStyle: InheritParentConfig
-DisableFormat: false
