+++ generator/integration_tests/google_cloud_cpp_generator_golden_tests.bzl
