+++ generator/integration_tests/tests/golden_kitchen_sink_client_test.cc
