+++ generator/integration_tests/tests/golden_kitchen_sink_connection_test.cc
@@ -20,7 +20,7 @@
 #include "generator/integration_tests/golden/golden_kitchen_sink_options.h"
 #include "generator/integration_tests/golden/internal/golden_kitchen_sink_connection_impl.h"
 #include "generator/integration_tests/golden/internal/golden_kitchen_sink_option_defaults.h"
-#include "generator/integration_tests/golden/mocks/mock_golden_kitchen_sink_stub.h"
+#include "generator/integration_tests/tests/mock_golden_kitchen_sink_stub.h"
 #include <gmock/gmock.h>
 #include <memory>
