+++ generator/integration_tests/tests/golden_kitchen_sink_idempotency_policy_test.cc
