+++ generator/integration_tests/tests/golden_kitchen_sink_option_defaults_test.cc
