+++ generator/integration_tests/tests/golden_kitchen_sink_stub_factory_test.cc
