+++ generator/integration_tests/tests/golden_kitchen_sink_stub_test.cc
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 #include "generator/integration_tests/golden/internal/golden_kitchen_sink_stub.h"
-#include "google/cloud/testing_util/status_matchers.h"
 #include "google/cloud/testing_util/mock_completion_queue_impl.h"
+#include "google/cloud/testing_util/status_matchers.h"
 #include "absl/memory/memory.h"
 #include <gmock/gmock.h>
 #include <grpcpp/impl/codegen/status_code_enum.h>
