+++ generator/integration_tests/tests/golden_thing_admin_client_test.cc
