+++ generator/integration_tests/tests/golden_thing_admin_idempotency_policy_test.cc
