+++ generator/integration_tests/tests/golden_thing_admin_option_defaults_test.cc
