+++ generator/integration_tests/tests/golden_thing_admin_stub_test.cc
