+++ generator/integration_tests/tests/mock_golden_thing_admin_stub.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_GENERATOR_INTEGRATION_TESTS_GOLDEN_MOCKS_MOCK_GOLDEN_THING_ADMIN_STUB_H
-#define GOOGLE_CLOUD_CPP_GENERATOR_INTEGRATION_TESTS_GOLDEN_MOCKS_MOCK_GOLDEN_THING_ADMIN_STUB_H
+#ifndef GOOGLE_CLOUD_CPP_GENERATOR_INTEGRATION_TESTS_TESTS_MOCK_GOLDEN_THING_ADMIN_STUB_H
+#define GOOGLE_CLOUD_CPP_GENERATOR_INTEGRATION_TESTS_TESTS_MOCK_GOLDEN_THING_ADMIN_STUB_H
 
 #include "generator/integration_tests/golden/internal/golden_thing_admin_stub.h"
 #include <gmock/gmock.h>
@@ -36,8 +36,7 @@ class MockGoldenThingAdminStub
       (override));
   MOCK_METHOD(
       future<StatusOr<::google::longrunning::Operation>>, AsyncCreateDatabase,
-      (google::cloud::CompletionQueue&,
-       std::unique_ptr<grpc::ClientContext>,
+      (google::cloud::CompletionQueue&, std::unique_ptr<grpc::ClientContext>,
        ::google::test::admin::database::v1::CreateDatabaseRequest const&),
       (override));
   MOCK_METHOD(StatusOr<::google::test::admin::database::v1::Database>,
@@ -46,9 +45,9 @@ class MockGoldenThingAdminStub
                ::google::test::admin::database::v1::GetDatabaseRequest const&),
               (override));
   MOCK_METHOD(
-      future<StatusOr<::google::longrunning::Operation>>, AsyncUpdateDatabaseDdl,
-      (google::cloud::CompletionQueue&,
-          std::unique_ptr<grpc::ClientContext>,
+      future<StatusOr<::google::longrunning::Operation>>,
+      AsyncUpdateDatabaseDdl,
+      (google::cloud::CompletionQueue&, std::unique_ptr<grpc::ClientContext>,
        ::google::test::admin::database::v1::UpdateDatabaseDdlRequest const&),
       (override));
   MOCK_METHOD(Status, DropDatabase,
@@ -74,9 +73,10 @@ class MockGoldenThingAdminStub
               (grpc::ClientContext&,
                ::google::iam::v1::TestIamPermissionsRequest const&),
               (override));
-  MOCK_METHOD(future<StatusOr<::google::longrunning::Operation>>, AsyncCreateBackup,
+  MOCK_METHOD(future<StatusOr<::google::longrunning::Operation>>,
+              AsyncCreateBackup,
               (google::cloud::CompletionQueue&,
-                  std::unique_ptr<grpc::ClientContext>,
+               std::unique_ptr<grpc::ClientContext>,
                ::google::test::admin::database::v1::CreateBackupRequest const&),
               (override));
   MOCK_METHOD(StatusOr<::google::test::admin::database::v1::Backup>, GetBackup,
@@ -100,8 +100,7 @@ class MockGoldenThingAdminStub
       (override));
   MOCK_METHOD(
       future<StatusOr<::google::longrunning::Operation>>, AsyncRestoreDatabase,
-      (google::cloud::CompletionQueue&,
-          std::unique_ptr<grpc::ClientContext>,
+      (google::cloud::CompletionQueue&, std::unique_ptr<grpc::ClientContext>,
        ::google::test::admin::database::v1::RestoreDatabaseRequest const&),
       (override));
   MOCK_METHOD(
@@ -121,36 +120,36 @@ class MockGoldenThingAdminStub
       (override));
 
   MOCK_METHOD(future<StatusOr<::google::test::admin::database::v1::Database>>,
-      AsyncGetDatabase,
+              AsyncGetDatabase,
               (google::cloud::CompletionQueue&,
-                  std::unique_ptr<grpc::ClientContext>,
+               std::unique_ptr<grpc::ClientContext>,
                ::google::test::admin::database::v1::GetDatabaseRequest const&),
               (override));
 
   MOCK_METHOD(future<Status>, AsyncDropDatabase,
               (google::cloud::CompletionQueue&,
-                  std::unique_ptr<grpc::ClientContext>,
+               std::unique_ptr<grpc::ClientContext>,
                ::google::test::admin::database::v1::DropDatabaseRequest const&),
               (override));
 
   MOCK_METHOD(
       future<StatusOr<::google::longrunning::Operation>>,
       AsyncLongRunningWithoutRouting,
-      (google::cloud::CompletionQueue&,
-          std::unique_ptr<grpc::ClientContext>,
+      (google::cloud::CompletionQueue&, std::unique_ptr<grpc::ClientContext>,
        ::google::test::admin::database::v1::RestoreDatabaseRequest const&),
       (override));
 
-  MOCK_METHOD(future<StatusOr<google::longrunning::Operation>>, AsyncGetOperation,
+  MOCK_METHOD(future<StatusOr<google::longrunning::Operation>>,
+              AsyncGetOperation,
               (google::cloud::CompletionQueue&,
-                  std::unique_ptr<grpc::ClientContext>,
+               std::unique_ptr<grpc::ClientContext>,
                google::longrunning::GetOperationRequest const&),
               (override));
 
   MOCK_METHOD(future<Status>, AsyncCancelOperation,
               (google::cloud::CompletionQueue&,
-                  std::unique_ptr<grpc::ClientContext>,
-                  google::longrunning::CancelOperationRequest const&),
+               std::unique_ptr<grpc::ClientContext>,
+               google::longrunning::CancelOperationRequest const&),
               (override));
 };
 
@@ -159,4 +158,4 @@ GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_GENERATOR_INTEGRATION_TESTS_GOLDEN_MOCKS_MOCK_GOLDEN_THING_ADMIN_STUB_H
+#endif  // GOOGLE_CLOUD_CPP_GENERATOR_INTEGRATION_TESTS_TESTS_MOCK_GOLDEN_THING_ADMIN_STUB_H
